Add function for checking if a path is a retail install
This checks if a path is a retail install by checking if root_fs2.vp
exists. Since this needs to be done case-insensitive (since FSO works
that way) I added a function that does this and used this function where
appropriate.

This fixes a bug for me where Knossos refused to use my FreeSpace
install since the root file was named "Root_fs2.vp".