Add mod.ini import support
This adds an action to import a legacy mod into the Knossos system which
performs the most annoying steps automatically. This will try to convert
the old style dependency management to the new style by using some
hard-coded conversion values. Currently, that table only contains the
various MediaVP versions.

This will also extract the individual VPs and create packages for each
of them.