Add mod action to verify file integrity
This completely replaces the previous CheckTask which caused some issues
if the file hashes were not found in the hash cache. This has been
replaced by a dedicated action which can be triggered by the user when
desired.

The action currently only validates the files and displays an error
dialog if any invalid files have been found. This could be extended to
automatically redownload the corrupted or missing files but that has not
been included here.