Add flag parsing using the new JSON format
I encountered an issue with the environment that the executable is
started with. If it only contained the LD_LIBRARY_PATH entry then the
SDL code within FSO didn't work for some reason. If the env parameter
was omitted then it worked properly so I emulated the default value by
copying the existing process environment and then change that in case
any changes are needed at all. This fixed the issue on my system.