Fix libffi crash when calling SDL_JoystickGetGUIDString
It appears like there is a known issue with passing a 16-byte array by
value to a C-function: https://bugs.python.org/issue22273

I fixed that by simply splitting the array into two smaller arrays which
fixed the error for me. SDL still returns the correct values so it seems
like it works for the C-code.