Also include mods with higher stability in mode lists
The stability setting only specifies what the minimum stability of the
engine must be but the code which builds the mod version list used the
value in an equality check. This fixes that by adding a new function to
Mod which checks if the mod is at least as stable as requested.