Use task for applying flags to all engines
For me the "Apply to all" button causes the main window to become
unresponsive since I have a lot of engine versions installed. Knossos
retrieves the flags in the main thread so until all engine versions have
been checked, the main window becomes unresponsive.

This fixes that by using a background task for retrieving the engine
flags which minimizes the work that is done on the main thread.