Skip the cmdline_fso.cfg writing when launching FSO
The original reason why that file was used was to allow users to
directly launch FSO from the file manager. Since that doesn't work
anyway in Knossos there is no reason to keep writing that file.

This is a minor change which makes it easier to use Knossos when
starting FSO using different launchers.

I added a save_fso_settings call before launching FSO to make sure that
the config file is updated before FSO is launched. This will make sure
that the engine uses the correct settings path.