Fix issues with packages with prerelease versions
It seems that the previous version of semantic-version did not look at
the prerelease part of the version when comparing them so Knossos
thought that 19.0.0-RC1 was the same as 19.0.0 and so did not add the
stable release to its internal mod list. The update of the package fixes
that.