Use parts lock for all accesses to the variables
In the MultipartUploader there was a member that contained the parts to
upload or something and that was read and written to in various places
without any locking which resulted in some inconsistent states and
subsequent exceptions. This should fix that by using the lock wherever
the variable is accessed.